# PetsShelterBack/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Pets Shelter
# during dev, should point to your local machine
host: localhost:3000
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /species:
    x-swagger-router-controller: specie.controller
    get:
      description: Get the species list
      operationId: getSpecies
      parameters:
        - name: name  ## param to filter
          in: query
          type: string
          required: false
        - name: sort  ## param to order
          in: query
          type: string
          required: false
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetSpecieListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Add a new specie to the list
      operationId: createSpecie
      parameters:
        - in: body
          name: specie
          description: The specie to be created.
          schema:
            $ref: "#/definitions/Specie"
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/GetSpecieResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /species/{id}:
    x-swagger-router-controller: specie.controller
    get:
      description: get a specie by id
      operationId: getSpecieById
      parameters:
        - name: id #important !!!, #name# is a property of swagger
          type: string #id is a string type
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetSpecieResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      description: update a specie
      operationId: updateSpecie
      parameters:
        - name: id  # the id to modify
          description: Specie's id
          type: string
          in: path
          required: true
        - in: body  #the modified data
          name: specie
          description: The specie to update.
          schema:
            $ref: "#/definitions/Specie"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetSpecieResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      description: delete a specie
      operationId: deleteSpecie
      parameters:
        - name: id
          description: Specie's id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /hello:
    # binds a127 app logic to a route
    x-swagger-router-controller: hello_world
    get:
      description: Returns 'Hello' to the caller
      # used as the method name of the controller
      operationId: hello
      parameters:
        - name: name
          in: query
          description: The name of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string

  Specie:
    type: object
    properties:
      name:
        type: string
        description: Name of the Specie
    required:
      - name

  GetSpecieResponse:
    type: object
    properties:
      id:
        type: string
        description: Identifier
      name:
        type: string
        description: Name of the Specie
    required:
      - name

  GetSpecieListResponse:
    properties:
      species:
        type: array
        description: Species List
        items:
          $ref: "#/definitions/GetSpecieResponse"
    required:
      - species

  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: a short comment
    required:
      - success
      - description

  ErrorResponse:
    properties:
      message:
        type: string
        description: Error message
    required:
      - message